dataverse-dev/requirements.txt
--trusted-host gitlab.renewbuy.in
--extra-index-url https://__token__:8zKkkbRtGyJ1LdxxUshA@gitlab.renewbuy.in/api/v4/projects/103/packages/pypi/simple
rb-utils==2023.2.7.7.59.17
aioboto3==10.2.0
aiobotocore==2.4.1
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.9.0
anyio==3.6.2
asyncpg==0.25.0
async-timeout==4.0.2
attrs==22.1.0
boto3==1.24.59
botocore==1.27.59
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
httpcore==0.13.2
httpx==0.18.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.2.0
importlib-resources==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.1
motor==2.5.0
multidict==6.0.3
numpy==1.24.0
pandas==1.5.2
pydantic==1.10.2
pymongo==3.12.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.7
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.2.1
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
wrapt==1.14.1
WTForms==3.0.1
yarl==1.8.2
zipp==3.11.0
fastapi-pagination==0.11.1
email-validator==1.2.1